requirements.txt
pyautogui==0.9.54
opencv-python==4.10.0.84
Pillow==10.4.0
